chore: refactor files' names
